fix SVG Blob feature detection
Found in #2

We stupidly never used the is_svg_blob result and thus ended up always using the polyfill for any Blob input, even in browsers that do natively support it.

This can't be caught by WPT so we really need to write tests for it and other areas specific only to the polyfill.